+++ public/assets/js/favorites.js
@@ -4,8 +4,9 @@ $(document).ready(function () {
     // Delete the favorite when the user clicks the delete button
     $(".delete-btn").click(function (event) {
         var id = $(event.target).data("id");
+
         $.ajax({
-            url: `/favorites/${id}`,
+            url: `/favorites/${id}/` + getCookie('currentUser'),
             method: "DELETE"
         }).done(function () {
             $(`#card-${id}`).remove()
@@ -14,3 +15,21 @@ $(document).ready(function () {
         })
     });
 });
+
+console.log(getCookie('currentUser'))
+
+function getCookie(cname) {
+    var name = cname + "=";
+    var decodedCookie = decodeURIComponent(document.cookie);
+    var ca = decodedCookie.split(';');
+    for(var i = 0; i <ca.length; i++) {
+      var c = ca[i];
+      while (c.charAt(0) == ' ') {
+        c = c.substring(1);
+      }
+      if (c.indexOf(name) == 0) {
+        return c.substring(name.length, c.length);
+      }
+    }
+    return "";
+  }
